src/account_service/requirements.txt
grpcio
grpcio-tools
pymongo
flask
thread6
pytest         # NEW
pytest-mock    # NEW